+++ packages/lit-html/src/directives/render-light.ts
@@ -14,13 +14,14 @@
 
 import {ChildPart} from '../lit-html.js';
 import {directive, Directive} from '../directive.js';
-import {isDirectiveResult} from '../directive-helpers.js';
+import {classForDirectiveResult} from '../directive-helpers.js';
 
 export interface RenderLightHost extends HTMLElement {
   renderLight(): unknown;
 }
 
 class RenderLight extends Directive {
+  static _$litRenderLight = true;
   render() {
     /* SSR handled specially in render-lit-html */
   }
@@ -119,4 +120,4 @@ class RenderLight extends Directive {
 export const renderLight = directive(RenderLight);
 
 export const isRenderLightDirective = (value: unknown): boolean =>
-  isDirectiveResult(value, RenderLight);
+  (classForDirectiveResult(value) as typeof RenderLight)?._$litRenderLight;